댓글 삭제 오류 해결

+++ board/src/main/java/DAO/CommentDAO.java
@@ -50,13 +50,13 @@ public class CommentDAO {
         }
     }
 
-    public static void postCommentDelete(int postId) throws Exception{
+    public static void postCommentDelete(long postId) throws Exception{
         String sql = "delete from comment " +
                      "where postId = ?";
 
         try (Connection connection = DB.getConnection();
              PreparedStatement statement = connection.prepareStatement(sql)) {
-             statement.setInt(1, postId);
+             statement.setLong(1, postId);
              statement.executeUpdate();
         }
     }
